Segregate old data access code.  #3496.
