+++ app/models/response_set_prepopulation/low_intensity.rb
@@ -1,17 +1,19 @@
 module ResponseSetPrepopulation
   class LowIntensity < Populator
+    include OldAccessMethods
+
+    def self.applies_to?(rs)
+      rs.survey.title.include?('_QUE_LI')
+    end
+
     def reference_identifiers
       [
         "prepopulated_ppg_status",
       ]
     end
-    
-    def self.applies_to?(rs)
-      rs.survey.title.include?('_QUE_LI')
-    end
 
     def run
-        reference_identifiers.each do |reference_identifier|
+      reference_identifiers.each do |reference_identifier|
         if question = find_question_for_reference_identifier(reference_identifier)
           answer = question.answers.first
           value = case reference_identifier
