+++ app/models/response_set_prepopulation/pregnancy_screener.rb
@@ -1,5 +1,7 @@
 module ResponseSetPrepopulation
   class PregnancyScreener < Populator
+    include OldAccessMethods
+
     def self.applies_to?(rs)
       rs.survey.title.include?('_PregScreen_')
     end
